pyproject.toml
[project]
name = "naoseionome"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Marco André Mendes", email = "marcoandre@gmail.com" },
    {name = "GustavoLopesMaximiano", email = "gustavo.maximiano.ifc@gmail.com"},
]
dependencies = [
    'dj-database-url>=2.1.0',
    'django-cloudinary-storage>=0.3.0',
    'django-cors-headers>=4.3.0',
    'django-extensions>=3.2.3',
    'django-filter>=23.3',
    "django>=5.1.3",
    'djangorestframework>=3.14.0',
    'drf-spectacular>=0.26.5',
    'gunicorn>=21.2.0',
    # 'mysqlclient>=2.2.4',
    'passage-identity>=2.5.1',
    'Pillow>=10.3.0',
    'pre-commit>=3.5.0',
    'psycopg2-binary>=2.9.9',
    'pydotplus>=2.0.2',
    'python-dotenv>=1.0.0',
    'python-magic-bin>=0.4.14; sys_platform == "win32"',
    'python-magic>=0.4.27; sys_platform == "linux"',
    'setuptools>=68.2.2',
    'uvicorn>=0.29.0',
    'whitenoise[brotli]>=6.6.0',
    "annotated-types==0.7.0",
    "asgiref==3.8.1",
    "astroid==3.3.9",
    "asttokens==3.0.0",
    "attrs==24.2.0",
    "brotli==1.1.0",
    "certifi==2024.8.30",
    "cffi==1.17.1; platform_python_implementation != \"PyPy\"",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "cloudinary==1.41.0",
    "colorama==0.4.6; sys_platform == \"win32\" or platform_system == \"Windows\"",
    "cryptography==44.0.2",
    "decorator==5.2.1",
    "dill==0.3.9",
    "distlib==0.3.9",
    "dj-database-url==2.3.0",
    "django==5.1.7",
    "django-cloudinary-storage==0.3.0",
    "django-cors-headers==4.7.0",
    "django-extensions==3.2.3",
    "django-filter==25.1",
    "djangorestframework==3.15.2",
    "drf-spectacular==0.28.0",
    "exceptiongroup==1.2.2; python_version < \"3.11\"",
    "executing==2.2.0",
    "filelock==3.16.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "identify==2.6.1",
    "idna==3.10",
    "inflection==0.5.1",
    "ipython==8.34.0",
    "isort==6.0.1",
    "jedi==0.19.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "matplotlib-inline==0.1.7",
    "mccabe==0.7.0",
    "mysqlclient==2.2.7",
    "netifaces==0.11.0",
    "nodeenv==1.9.1",
    "packaging==24.1",
    "parso==0.8.4",
    "passage-identity==3.0.1",
    "pexpect==4.9.0; sys_platform != \"win32\" and sys_platform != \"emscripten\"",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "pre-commit==4.2.0",
    "prompt-toolkit==3.0.50",
    "psycopg2-binary==2.9.10",
    "ptyprocess==0.7.0; sys_platform != \"win32\" and sys_platform != \"emscripten\"",
    "pure-eval==0.2.3",
    "pycparser==2.22; platform_python_implementation != \"PyPy\"",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydotplus==2.0.2",
    "pygments==2.19.1",
    "pyjwt==2.9.0",
    "pylint==3.3.6",
    "pyparsing==3.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-magic==0.4.27; sys_platform == \"linux\"",
    "python-magic-bin==0.4.14; sys_platform == \"win32\"",
    "pyyaml==6.0.2",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rpds-py==0.20.1",
    "ruff==0.11.2",
    "setuptools==78.0.2",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "stack-data==0.6.3",
    "tomli==2.0.2; python_version < \"3.11\"",
    "tomlkit==0.13.2",
    "traitlets==5.14.3",
    "typing-extensions==4.12.2",
    "tzdata==2024.2; sys_platform == \"win32\"",
    "uritemplate==4.1.1",
    "urllib3==2.0.7",
    "uvicorn==0.34.0",
    "virtualenv==20.27.1",
    "wcwidth==0.2.13",
    "whitenoise[brotli]==6.9.0",
]
requires-python = ">=3.13"
license = { text = "MIT" }
readme = "README.md"

[dependency-groups]
dev = [
  "pylint>=3.0.2",
  "ipython>=8.26.0",
  "ruff>=0.11.2",
]

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 127.0.0.1:1800"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
post_migrate = "python manage.py graph_models -S -g -o core.png core"
shell = "python manage.py shell"
shellp = "python manage.py shell_plus"
test = "python manage.py test"
loaddata = "python manage.py loaddata core.json"
dumpdata = "python manage.py dumpdata --indent 2"
cria_api = "python ./scripts/cria_api.py {args}"
check = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"

[tool.pdm]
distribution = true

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.ruff]
line-length = 120
extend-exclude = ['migrations', '__init__.py', 'manage.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["PLR6301"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
